#93 문의하기 기능 옮기기 Detail -> Info
